refactor: restructure `action.rs` into `action/`
