+++ src-tauri/identity_wallet/src/state/actions/cancel_user_journey.rs
@@ -1,0 +1,17 @@
+use crate::reducer;
+use crate::state::actions::ActionTrait;
+use crate::state::actions::Reducer;
+use crate::state::reducers::cancel_user_journey;
+
+use ts_rs::TS;
+
+#[derive(serde::Serialize, serde::Deserialize, Debug, TS, Clone)]
+#[ts(export, export_to = "bindings/actions/CancelUserJourney.ts")]
+pub struct CancelUserJourney;
+
+#[typetag::serde(name = "[User Journey] Cancel")]
+impl ActionTrait for CancelUserJourney {
+    fn reducers<'a>(&self) -> Vec<Reducer<'a>> {
+        vec![reducer!(cancel_user_journey)]
+    }
+}
