+++ src-tauri/identity_wallet/src/state/actions/credentials_selected.rs
@@ -1,0 +1,20 @@
+use crate::reducer;
+use crate::state::actions::ActionTrait;
+use crate::state::actions::Reducer;
+use crate::state::reducers::authorization::handle_oid4vp_authorization_request;
+
+use ts_rs::TS;
+
+#[derive(serde::Serialize, serde::Deserialize, Debug, TS, Clone)]
+#[ts(export, export_to = "bindings/actions/CredentialsSelected.ts")]
+pub struct CredentialsSelected {
+    #[ts(type = "Array<string>")]
+    pub credential_uuids: Vec<uuid::Uuid>,
+}
+
+#[typetag::serde(name = "[Authenticate] Credentials selected")]
+impl ActionTrait for CredentialsSelected {
+    fn reducers<'a>(&self) -> Vec<Reducer<'a>> {
+        vec![reducer!(handle_oid4vp_authorization_request)]
+    }
+}
