+++ src-tauri/identity_wallet/src/state/actions/reset.rs
@@ -1,0 +1,17 @@
+use crate::reducer;
+use crate::state::actions::ActionTrait;
+use crate::state::actions::Reducer;
+use crate::state::reducers::reset_state;
+
+use ts_rs::TS;
+
+#[derive(serde::Serialize, serde::Deserialize, Debug, TS, Clone)]
+#[ts(export, export_to = "bindings/actions/Reset.ts")]
+pub struct Reset;
+
+#[typetag::serde(name = "[App] Reset")]
+impl ActionTrait for Reset {
+    fn reducers<'a>(&self) -> Vec<Reducer<'a>> {
+        vec![reducer!(reset_state)]
+    }
+}
