+++ src-tauri/identity_wallet/src/state/actions/set_locale.rs
@@ -1,0 +1,21 @@
+use crate::reducer;
+use crate::state::actions::ActionTrait;
+use crate::state::actions::Reducer;
+use crate::state::reducers::set_locale;
+use crate::state::Locale;
+
+use ts_rs::TS;
+
+#[derive(serde::Serialize, serde::Deserialize, Debug, TS, Clone)]
+#[ts(export, export_to = "bindings/actions/SetLocale.ts")]
+pub struct SetLocale {
+    #[ts(type = "string")]
+    pub locale: Locale,
+}
+
+#[typetag::serde(name = "[Settings] Set locale")]
+impl ActionTrait for SetLocale {
+    fn reducers<'a>(&self) -> Vec<Reducer<'a>> {
+        vec![reducer!(set_locale)]
+    }
+}
